Tiberium Alliances - xTr1m's Base Overlay v1.0
Provides an overlay over the base of the gain/cost relationship of building upgrades by holding the Control key while in base view (weird coloured squares with numbers otherwise).

It's a direct implementation of CNCOpt's "Figuring out which buildings to upgrade first" function: "This will put an overlay of what buildings give you the highest increase in production per upgrade cost. The number displayed is the increase in production / upgrade cost ratio, so the higher the better. The highest values will be highlighted in green, so generally speaking upgrade the green buildings first."

Author: https://github.com/xTr1m/